update authenticated vs unauth page structure
