+++ src/components/DisplayNews.jsx
@@ -6,30 +6,20 @@ import Footer from "./Footer";
 import Weather from "./Weather"
 import Headlines from "./Headlines"
 import Trending from "./Trending"
+import Featured from "./Featured"
 import { connect } from "react-redux";
-import { getFreeArticle, getArticles } from "../modules/article";
+import { getFreeArticle } from "../modules/article";
 
 const DisplayNews = props => {
   const [freeContent, setFreeContent] = useState(null)
-  const [featured, setFeatured] = useState(null)
 
   const loadFreeArticle = async () => {
     let response = await getFreeArticle();
     setFreeContent(response)
   }
 
-  const loadFeatured =  async () => {
-    let response = await getArticles(
-      props.language,
-      1,
-      "sports"
-    );
-    setFeatured(response)
-  }
-
   useEffect(() => {
     loadFreeArticle();
-    loadFeatured()
   }, []);
   
   return (
@@ -45,68 +35,8 @@ const DisplayNews = props => {
           </div>
           {props.authenticated === false && (
             <>
-              <div id="seperator" className="item-c">
-                <div id="border-top" className="feature">
-                  {featured !== null && featured.articles.length > 0 && (
-                    <>
-                      <h4 
-                        onClick={() => props.changeCurrentArticleId(featured.articles[0].id)} 
-                        id="featured-title" 
-                      >
-                        {featured.articles[0].title}
-                      </h4>
-                      <img id="featured-img" src={featured.articles[0].image} />
-                      <h5 id="featured-body">{featured.articles[0].body}</h5>
-                      <p id="cat-date" >
-                        <span id="red">{featured.articles[0].category} </span> 
-                        {featured.articles[0].created_at.substring(0, featured.articles[0].created_at.lastIndexOf("T"))}
-                      </p>
-                      <p id="feature-hover">FEATURED</p>
-                    </>
-                  )}
-                </div>
-              </div>
-              <div id="seperator" className="item-d">
-                <div id="border-top" className="feature">
-                  {featured !== null && featured.articles.length > 1 && (
-                    <>
-                      <h4 
-                        onClick={() => props.changeCurrentArticleId(featured.articles[1].id)} 
-                        id="featured-title" 
-                      >
-                        {featured.articles[1].title}
-                      </h4>
-                      <img id="featured-img" src={featured.articles[1].image} />
-                      <h5 id="featured-body">{featured.articles[1].body}</h5>
-                      <p id="cat-date" >
-                        <span id="red">{featured.articles[1].category} </span> 
-                        {featured.articles[1].created_at.substring(0, featured.articles[1].created_at.lastIndexOf("T"))}
-                      </p>
-                      <p id="feature-hover">FEATURED</p>
-                    </>
-                  )}
-                </div>
-              </div>
-              <div className="item-e">
-                <div id="border-top" className="feature">
-                  {featured !== null && featured.articles.length > 2 && (
-                    <>
-                      <h4 
-                        onClick={() => props.changeCurrentArticleId(featured.articles[2].id)} 
-                        id="featured-title" 
-                      >
-                        {featured.articles[2].title}
-                      </h4>
-                      <img id="featured-img" src={featured.articles[2].image} />
-                      <h5 id="featured-body">{featured.articles[2].body}</h5>
-                      <p id="cat-date" >
-                        <span id="red">{featured.articles[2].category} </span> 
-                        {featured.articles[2].created_at.substring(0, featured.articles[2].created_at.lastIndexOf("T"))}
-                      </p>
-                      <p id="feature-hover">FEATURED</p>
-                    </>
-                  )}
-                </div>
+              <div className="item-c">
+                <Featured />
               </div>
             </>
           )}
@@ -129,38 +59,51 @@ const DisplayNews = props => {
           </div>
         </div>
       </div>
-      {props.authenticated === false && (
-        <div className="whitebackground">
-          <div className="main2">
-            <div className="item-h">
-                <div id="border-top">
-                  {freeContent !== null && (
-                    <>
-                      <h2>FREE READ</h2>
-                      <img className="free-read-img" src={freeContent[1]} />
-                    </>
-                  )}
-                </div>
-            </div>
-            <div className="item-i">
-              {freeContent !== null && freeContent.length > 0 && (
-                <div className="free-read">
-                  <h1>{freeContent[0].title}</h1>
-                  <p>{freeContent[0].body}</p>
-                  <p id="cat-date" >
-                    <span id="red">{freeContent[0].category}</span> {freeContent[0].created_at.substring(0, freeContent[0].created_at.lastIndexOf("T"))}
-                  </p>
-                </div>
-              )}
-            </div>
-            <div className="item-g">
-              <div className="advertising">
-                ADVERTISING
+      <div className="whitebackground">
+        <div className="main2">
+          {props.authenticated === false ? (
+            <>
+              <div className="item-h">
+                  <div id="border-top">
+                    {freeContent !== null && (
+                      <>
+                        <h2>FREE READ</h2>
+                        <img className="free-read-img" src={freeContent[1]} />
+                      </>
+                    )}
+                  </div>
               </div>
-          </div>
-          </div>
+              <div className="item-i">
+                {freeContent !== null && freeContent.length > 0 && (
+                  <div className="free-read">
+                    <h1>{freeContent[0].title}</h1>
+                    <p>{freeContent[0].body}</p>
+                    <p id="cat-date" >
+                      <span id="red">{freeContent[0].category}</span> {freeContent[0].created_at.substring(0, freeContent[0].created_at.lastIndexOf("T"))}
+                    </p>
+                  </div>
+                )}
+              </div>
+              <div className="item-g">
+                <div className="advertising">
+                  ADVERTISING
+                </div>
+              </div>
+            </>
+          ) : (
+            <>
+              <div className="item-d">
+                <Featured />
+              </div>
+              <div className="item-e">
+                <div className="advertising">
+                  ADVERTISING
+                </div>
+              </div>
+            </>
+          )}
         </div>
-      )}
+      </div>
       <Footer />
     </>
   );
@@ -173,15 +116,6 @@ const mapStateToProps = state => {
   }
 }
 
-const mapDispatchToProps = dispatch => {
-  return {
-    changeCurrentArticleId: id => {
-      dispatch({ type: "CHANGE_ARTICLE_ID", payload: id });
-    }
-  };
-};
-
 export default connect(
-  mapStateToProps,
-  mapDispatchToProps
+  mapStateToProps
 )(DisplayNews);
