+++ src/index.css
@@ -40,12 +40,14 @@ html {
 }
 
 .footer-container {
+  min-width: 1245px;
   background-color: #2222228c;
 }
 
 .backgroundgrey {
   background-color: #EDEDED;
   padding-top: 60px;
+  min-width: 1245px;
 }
 
 .whitebackground {
@@ -71,7 +73,6 @@ html {
   font-family: "ChomskyRegular";
   font-weight: 900;
   float: right;
-  /* line-height: 140px; */
   font-size: 16px;
 }
 
@@ -110,6 +111,11 @@ a {
   font-size: larger;
   font-weight: bold;
   padding-bottom: 14px;
+}
+
+.create-header {
+  height: 60px;
+  font-weight: 100!important;
 }
 
 #stripe-form {
@@ -175,6 +181,7 @@ a#back-button.signup-back {
 
 .nav {
   background-color: rgba(249, 249, 249, 0.715);
+  min-width: 1245px;
 }
 
 .nav-bar {
@@ -196,14 +203,24 @@ a#back-button.signup-back {
   margin: 0 auto;
   display: grid;
   grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-template-rows: 600px auto;
+  grid-template-rows: auto auto;
   grid-row-gap: 15px;
 }
 
-#seperator {
+#seperator-a {
   border-right: 1px solid lightgrey;
   height: 90%;
-  margin-top: 5%;
+  padding-right: 5px;
+}
+
+#seperator-b {
+  border-right: 1px solid lightgrey;
+  height: 90%;
+  padding: 0 5px;
+}
+
+#seperator-c {
+  padding-left: 5px;
 }
 
 #border-top {
@@ -230,28 +247,18 @@ a#back-button.signup-back {
 }
 
 .item-c {
-  padding-right: 5px;
+  margin-top: 10px;
   grid-column-start: 1;
-  grid-column-end: span 1;
+  grid-column-end: span 3;
   grid-row-start: 2;
   grid-row-end: span 1;
 }
 
 .item-d {
-  padding-right: 5px;
-  padding-left: 5px;
-  grid-column-start: 2;
-  grid-column-end: span 1;
-  grid-row-start: 2;
-  grid-row-end: span 1;
-}
-
-.item-e {
-  margin-top: 5%;
-  padding-left: 5px;
-  grid-column-start: 3;
-  grid-column-end: span 1;
-  grid-row-start: 2;
+  margin-top: 10px;
+  grid-column-start: 1;
+  grid-column-end: span 3;
+  grid-row-start: 1;
   grid-row-end: span 1;
 }
 
@@ -263,6 +270,15 @@ a#back-button.signup-back {
 }
 
 .item-g {
+  margin-top: -22px;
+  padding-left: 5%;
+  grid-column-start: 4;
+  grid-column-end: span 1;
+  grid-row-start: 1;
+  grid-row-end: span 1;
+}
+
+.item-e {
   margin-top: -22px;
   padding-left: 5%;
   grid-column-start: 4;
@@ -457,6 +473,11 @@ h2#article-title {
   position: relative;
 }
 
+.feature2 {
+  padding: 5px 10px;
+  position: relative;
+}
+
 .feature:hover {
   background-color: rgb(228, 227, 227);
   height: 410px;
@@ -464,11 +485,19 @@ h2#article-title {
   transition: transform .5s ease;
 }
 
-.feature:hover #feature-hover {
+.feature2:hover {
+  background-color: rgb(247, 243, 243);
+  height: 410px;
+  transform: scale(1.04);
+  transition: transform .5s ease;
+}
+
+.feature:hover #feature-hover, 
+.feature2:hover #feature2-hover {
   display: block;
 }
 
-#feature-hover {
+#feature-hover, #feature2-hover  {
   display: none;
   position: absolute;
   left: 50%;
@@ -484,7 +513,13 @@ h2#article-title {
 
 .feature:hover #featured-img {
   filter: blur(2px);
-  transition: filter 0.5s ease;
+  /* transition: filter 0.5s ease; */
+}
+
+.featured-div {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  grid-template-rows: 400px;
 }
 
 #border-top h5 {
@@ -718,7 +753,7 @@ button {
 
 #heart {
   margin-left: 40px;
-  margin-top: -8px;
+  margin-top: -7px;
   font-size: 20px;
 }
 
@@ -862,8 +897,8 @@ button {
 #fade-in {
   position: fixed;
   top: -40px;
-  right: 55px;
-  width: 300px;
+  right: 100px;
+  max-width: 400px;
   height: 40px;
   color: rgb(81, 80, 80);
   background-color: #dfc38b;
@@ -898,4 +933,18 @@ button {
 
 @keyframes slideOut {
   0% { top: 0; }
+}
+
+#image img {
+  max-width: 400px;
+  max-height: 300px;
+}
+
+#user-comments {
+  padding: 2% 1%;
+  width: 95%;
+  /* padding-bottom: 20px; */
+  border-bottom: 1px solid grey;
+  margin-bottom: 40px;
+  background: rgb(250, 250, 250);
 }
